+++ source/test-suite/fixtures/StandaloneDerivingExpected.hs
@@ -1,0 +1,9 @@
+{-# LANGUAGE DataKinds #-}
+{-# LANGUAGE DerivingVia #-}
+{-# LANGUAGE StandaloneDeriving #-}
+module StandaloneDeriving where
+
+deriving via
+  (Generically (PtPostProcessConfig c))
+  instance
+    (Semigroup (c 'Partial)) => Semigroup (PtPostProcessConfig c)