+++ source/test-suite/fixtures/StandaloneDerivingInvalid.hs
@@ -1,0 +1,5 @@
+{-# LANGUAGE DerivingStrategies #-}
+{-# LANGUAGE StandaloneDeriving #-}
+module StandaloneDerivingInvalid where
+
+deriving stock instance