+++ source/test-suite/fixtures/StandaloneDerivingUnsupported.hs
@@ -1,0 +1,6 @@
+{-# LANGUAGE DataKinds #-}
+{-# LANGUAGE DerivingVia #-}
+{-# LANGUAGE StandaloneDeriving #-}
+module StandaloneDerivingUnsupported where
+
+deriving via '(Int, Bool) instance Show (Wrapper a)